Settings: Add setting in security for sms message rate alerts.
Currently this is null so it defaults to 30 sms messages in a 1 min period.
Creating and writing this preference allows us to control the sms rate interface.
We can also modify timeframe as well but that'll be later maybe.

This was added to Misc instead of sms or security since this isn't just rate limiting the
sms application.  This rate limits any application that sends texts.

This also creates a new category called Application Security.

@ashoksoni edits:
Added retrieving system values to show on onCreate else it was showing default values always.
Also retrieve current values when clicking on listpreference.

Change-Id: I4abb71011aa79e9346401b2f2255f380109740fa
Signed-off-by: Ashok Soni <ashok.soni@gmail.com>